Factor out wasm_opt testing. NFC.
Also, factor out auto-updating of binaryenjs testing so it lives
alongside the actual test code.
